Break into a separate pomodoro file and add a logo screen.
