+++ keyboards/lily58/keymaps/pomodoro/rules.mk
@@ -20,3 +20,4 @@ SRC +=  ./lib/rgb_state_reader.c \
         ./lib/host_led_state_reader.c \
         ./lib/timelogger.c \
         ./lib/keylogger.c \
+        pomodoro.c \
